extract ConsentEntityConverter
