+++ dpc-common/src/main/java/gov/cms/dpc/common/consent/entities/ConsentEntity.java
@@ -9,10 +9,17 @@ import javax.validation.constraints.NotNull;
 import java.io.Serializable;
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
+import java.time.ZoneId;
+import java.util.Optional;
 import java.util.UUID;
 
 @Entity(name = "consent")
 public class ConsentEntity implements Serializable {
+    public static final String LOINC_CATEGORY = "64292-6";
+    public static final String OPT_IN = "OPTIN";
+    public static final String OPT_OUT = "OPTOUT";
+    public static final String TREATMENT = "TREAT";
+    public static final String SCOPE_CODE = "patient-privacy";
 
     private static final long serialVersionUID = 8702499693412507926L;
 
@@ -128,5 +135,26 @@ public class ConsentEntity implements Serializable {
 
     public void setUpdatedAt(OffsetDateTime updatedAt) {
         this.updatedAt = updatedAt;
+    }
+
+    public static ConsentEntity defaultConsentEntity(Optional<UUID> id, Optional<String> hicn, Optional<String> mbi) {
+        ConsentEntity ce = new ConsentEntity();
+
+        ce.setId(UUID.randomUUID());
+        id.ifPresent(ce::setId);
+
+        ce.setCreatedAt(OffsetDateTime.now(ZoneId.of("UTC")));
+        ce.setEffectiveDate(LocalDate.now(ZoneId.of("UTC")));
+        ce.setUpdatedAt(OffsetDateTime.now(ZoneId.of("UTC")));
+
+        hicn.ifPresent(ce::setHicn);
+        mbi.ifPresent(ce::setMbi);
+
+        ce.setLoincCode(LOINC_CATEGORY);
+        ce.setPolicyCode(OPT_IN);
+        ce.setPurposeCode(TREATMENT);
+        ce.setScopeCode(SCOPE_CODE);
+
+        return ce;
     }
 }
